requirements.txt
aiosqlite==0.19.0
python-dotenv==1.0.1
pyserial==3.5
pynmea2==1.18.0
smbus2==0.4.3
dbus-next==0.2.3
gpiozero==2.0
pyalsaaudio==0.11.0
numpy==1.26.4
aiohttp==3.9.5



